docs: moves `Validation` section from More Examples to Number Mask
